+++ react/ui-components/src/atoms/stories/LabelFieldPair.stories.js
@@ -1,30 +1,35 @@
 import React from "react";
 import LabelFieldPair from "../LabelFieldPair";
+import TextBlock from "../TextBlock";
+import TextInput from "../TextInput";
 
 export default {
   title: "Atoms/LabelFieldPair",
   component: LabelFieldPair,
   argTypes: {
-    children: {
-      control: { type: "text" },
-    },
     className: {
-      control: { type: "text" },
+      control: "text",
     },
     style: {
       control: { type: "object" },
     },
+    vertical: {
+      control: "boolean",
+    },
   },
 };
 
-const Template = (args) => <LabelFieldPair {...args} />;
 
-export const Default = Template.bind({});
-Default.args = {
-  children: (
-    <>
-      <label>Field Label:</label>
-      <input type="text" placeholder="Enter value" />
-    </>
-  ),
-};
+export const Default = () => (
+  <LabelFieldPair>
+    <TextBlock body={"Name"}></TextBlock>
+    <TextInput type="text"></TextInput>
+  </LabelFieldPair>
+);
+
+export const VerticalAlignment = () => (
+  <LabelFieldPair vertical={true}>
+    <TextBlock body={"Name"}></TextBlock>
+    <TextInput type="text"></TextInput>
+  </LabelFieldPair>
+);