+++ react/ui-components/CHANGELOG.md
@@ -4,6 +4,53 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2-beta.13] - 2024-07-08
+### New Changes
+- Updated Button OnClick and added digit footer images 
+
+## [0.0.2-beta.12] - 2024-07-05
+### New Changes
+- Updated Button
+
+
+## [0.0.2-beta.11] - 2024-07-04
+### New Changes
+- Added ActionBar, Header, ActionButton variant for Button
+- Updated Timeline Molecule 
+
+## [0.0.2-beta.10] - 2024-06-28
+### New Changes
+- Added Card Variants 
+
+## [0.0.2-beta.9] - 2024-06-25
+### New Changes
+- Updated Uploader preview 
+
+## [0.0.2-beta.8] - 2024-06-24
+### New Changes
+- Added Title prop for Button
+
+## [0.0.2-beta.7] - 2024-06-24
+### New Changes
+- Added ViewMore button for Timeline Molecule. 
+
+## [0.0.2-beta.6] - 2024-06-19
+### New Changes
+- Added BreadCrumb Component. 
+
+## [0.0.2-beta.5] - 2024-06-17
+### New Changes
+- Added Timeline Molecule. 
+
+## [0.0.2-beta.4] - 2024-06-15
+### Changed
+- fix: Fixed prop for the document upload.
+
+## [0.0.2-beta.3] - 2024-06-14
+### Changed
+- Updated Storybook
+- Modified BackButton as BackLink
+
 ## [0.0.2-beta.1] - 2024-06-04
 ### Changed
 - Updated Popup Classname